requirements.txt
ipython==8.14.0
ipywidgets==8.1.0
matplotlib==3.7.2
numpy==1.24.3
opencv_python==4.8.0.74
pandas==2.0.3
Pillow==9.5.0
scikit_learn==1.3.0
seaborn==0.12.2
streamlit==1.25.0
tensorflow==2.13.0
tensorflow_intel==2.13.0
tqdm==4.65.0
